moved email service to submodule
