+++ app.js
@@ -15,7 +15,7 @@ const {Application} = require("./services/application");
 const {MongoQueries} = require("./crdc-datahub-database-drivers/mongo-queries");
 const {DatabaseConnector} = require("./crdc-datahub-database-drivers/database-connector");
 const {getCurrentTime} = require("./crdc-datahub-database-drivers/utility/time-utility");
-const {EmailService} = require("./services/email");
+const {EmailService} = require("./crdc-datahub-database-drivers/services/email");
 const {NotifyUser} = require("./services/notify-user");
 const {User} = require("./crdc-datahub-database-drivers/services/user");
 const {extractAndJoinFields} = require("./utility/string-util");
