--- resources/email-templates/notification-template.html
@@ -1,27 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <style>
-      body,
-      pre {
-        font-family: -apple-system, BlinkMacSystemFont, "Segoe UI", Roboto,
-          Oxygen, Ubuntu, Cantarell, "Open Sans", "Helvetica Neue", sans-serif;
-      }
-    </style>
-  </head>
-  <body>
-    <p>Dear {{ firstName }},</p>
-    <p>{{ message }}</p>
-      {{#if users}}
-      <ul>
-        {{#each users}}
-        <li>{{{ this }}}</li>
-        {{/each}}
-      </ul>
-      {{/if}}
-    <p>Sincerely,</p>
-    <p>CRDC Data Hub Admin Team</p>
-  </body>
-</html>
