+++ test/graphql-queries/batch-job.test.js
@@ -4,7 +4,7 @@ const {TEST_APPLICATION} = require("../test-constants");
 const {MongoQueries} = require("../../crdc-datahub-database-drivers/mongo-queries");
 const config = require("../../config");
 const {DATABASE_NAME} = require("../../crdc-datahub-database-drivers/database-constants");
-const {EmailService} = require("../../services/email");
+const {EmailService} = require("../../crdc-datahub-database-drivers/services/email");
 const {NotifyUser} = require("../../services/notify-user");
 const {User} = require("../../crdc-datahub-database-drivers/services/user");
 const {ApprovedStudiesService} = require("../../services/approved-studies");
